Split UI bundle by route with lazy-loaded chunks
